Remove unnecessary IDE files and refactor graph related code
Deleted IDE related files (.idea) that are not necessary for the project overall. Also refactored the graph related (.go and .templ) files to improve functionally, mainly around the usage of the UUID for ID generation and changes in the indicator classes. Additionally, updated package.json to include a new dependency.